Use 2-byte magical number
As seen in PS2 files these repeating number are 2-byte sized values opposing to
previously considered 1-byte.

+++ src/stx/headers/image.h
@@ -11,7 +11,7 @@ namespace stx {
 
   struct Image {
     bool has_e6;
-    uint8_t magical_number;
+    uint16_t magical_number;
     Geometry geometry;
     unsigned char* image_data;
   };
